requirements.txt
# Core — always needed
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
pyyaml>=6.0
requests>=2.28.0

# Boosting models (recommended)
xgboost>=1.7.0
lightgbm>=3.3.0

# Data loading
kaggle>=1.5.0
datasets>=2.0.0

# Deployment
fastapi>=0.95.0
uvicorn>=0.20.0

# Databases
psycopg2-binary>=2.9.0
pymongo>=4.0.0
sqlalchemy>=2.0.0

# Cloud & Streaming
boto3>=1.26.0
kafka-python>=2.0.0

# Images
Pillow>=9.0.0
